Pending system implemented
Added preview of the last 10 quotations

User now has to wait for admin to approve quotation.

Only after a quotation is approved, it can be downloaded.

If an admin submits a quotation, it is automatically approved and downloadable.

If an admin rejects a quotation, it is deleted from database

Added new TODOs